refactor: 중복타입 제거 , cSpell 추가
