+++ .vscode/settings.json
@@ -1,10 +1,15 @@
 {
   "cSpell.words": [
+    "appkey",
+    "autoload",
     "Chakra",
     "CHOCO",
+    "clusterer",
+    "dapi",
     "GEEN",
     "heys",
     "Iget",
+    "infowindow",
     "Ipost",
     "kakao",
     "marketstatus",
