+++ components/Admin/marketItem.tsx
@@ -12,7 +12,7 @@ import Link from 'next/link';
 import { useState } from 'react';
 
 import { patchMarketStatus } from '@/components/Api/Market';
-import { IMarketItem, MarketStatus } from '@/types/Admin';
+import { MarketItemProps, MarketStatus } from '@/types/Admin';
 import { numberWithHyphenMarket } from '@/utils/numberWithHyphen';
 
 import ApprovedContainer from './buttonContainer/approvedContainer';
@@ -26,7 +26,7 @@ export default function MarketItem({
   marketName,
   businessNumber,
   createdAt,
-}: IMarketItem) {
+}: MarketItemProps) {
   const [isDeleted, setIsDeleted] = useState(false);
   const queryClient = useQueryClient();
