+++ components/Admin/marketList.tsx
@@ -5,12 +5,12 @@ import { useInView } from 'react-intersection-observer';
 
 import { postMarketList } from '@/components/Api/Market';
 import ErrorPage from '@/components/Error';
-import { IMarketItem, MarketList } from '@/types/Admin';
+import { MarketItemProps, MarketListProps } from '@/types/Admin';
 
 import AdminListSkeleton from './adminLIstSkeleton';
 import MarketItem from './marketItem';
 
-export default function MarketList({ category }: MarketList) {
+export default function MarketList({ category }: MarketListProps) {
   const { ref, inView } = useInView();
   const { data, status, fetchNextPage, isFetchingNextPage } = useInfiniteQuery(
     ['승인 마켓 리스트', category],
@@ -38,7 +38,7 @@ export default function MarketList({ category }: MarketList) {
     <>
       <Grid gap={0}>
         {data?.pages.map((page) =>
-          page?.enrollments.map((item: IMarketItem) => (
+          page?.enrollments.map((item: MarketItemProps) => (
             <Stack key={item.enrollmentId}>
               <MarketItem
                 phoneNumber={item.phoneNumber}
