+++ components/Shared/headerTitle.tsx
@@ -2,9 +2,9 @@ import { Box, Flex, Text } from '@chakra-ui/react';
 import { useRouter } from 'next/router';
 
 import { ArrowBackIcon } from '@/public/icon';
-import { MarketTitle } from '@/types/Market';
+import { HeaderTitleProps } from '@/types/Shared';
 
-export default function HeaderTitle({ title }: MarketTitle) {
+export default function HeaderTitle({ title }: HeaderTitleProps) {
   const router = useRouter();
   return (
     <Flex h={14} p={4} justifyContent="space-between" alignItems="center">
