--- types/Market.ts
@@ -1,3 +1,0 @@
-export interface MarketTitle {
-  title: string;
-}
