+++ types/Shared.ts
@@ -2,3 +2,7 @@ export interface KakaoMapsProps {
   address: string;
   title: string;
 }
+
+export interface HeaderTitleProps {
+  title: string;
+}
